requirements.txt
tqdm
rotary_embedding_torch
transformers
tokenizers
inflect
progressbar
einops
unidecode
scipy
librosa
numba
ffmpeg
numpy
numba
torchaudio
threadpoolctl
llvmlite
appdirs